+++ src/helpers/Game.js
@@ -2,8 +2,6 @@ const { roleNames, groupNames } = require("./Role");
 
 exports.gamePhases = ["day", "night"];
 
-exports.gameStatuses = ["assigning-roles", "on-going", "done"];
+exports.gameStatuses = ["assigning-roles", "playing", "done"];
 
 exports.waitingForPossibilities = [...roleNames, ...groupNames, "mayor"];
-
-exports.gameActions = ["eat", "use-potion", "look", "protect", "shoot", "mark", "delegate", "settle", "vote"];