+++ src/helpers/Player.js
@@ -1,1 +1,3 @@
 exports.playerAttributes = ["mayor", "eaten", "drank-life-potion", "drank-death-potion", "protected", "raven-marked"];
+
+exports.playerActions = ["eat", "use-potion", "look", "protect", "shoot", "mark", "delegate", "settle", "vote"];